+++ src/app/components/featured-products/featured-products.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { HttpService } from 'src/app/services/http.service';
+import { DeletePopupComponent } from '../modals/delete-popup/delete-popup.component';
 
 @Component({
   selector: 'app-featured-products',
@@ -9,7 +11,9 @@ import { HttpService } from 'src/app/services/http.service';
 export class FeaturedProductsComponent implements OnInit {
   public featuredProductsData: FeaturedProductsComponent[] | any;
   public isresponsed: boolean = false;
-  constructor(private httpService: HttpService) {
+  public dialogValue: string | any;
+
+  constructor(private httpService: HttpService, public dialog: MatDialog) {
     this.httpService
       .get('products/featured')
       .then((data) => {
@@ -24,9 +28,22 @@ export class FeaturedProductsComponent implements OnInit {
   ngOnInit(): void {}
 
   deleteFeatured(id: string) {
-    this.httpService.delete(id, 'products/featured/remove').then((data) => {
-      console.log(data);
-      this.ngOnInit();
+    const dialogRef = this.dialog.open(DeletePopupComponent, {
+      data: `Are you sure you want to delete?`,
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      console.log(result.data);
+      this.dialogValue = result.data;
+
+      if (this.dialogValue == 'You Confirmed') {
+        this.httpService.delete(id, 'products/featured/remove').then((data) => {
+          console.log(data);
+          this.ngOnInit();
+        });
+      } else {
+        console.log('You dont wanted to delete this');
+      }
     });
   }
 }
